Add kotlin-annotations to PermissionController
Soong normally adds kotlin-annotations as a dependency to any module
that uses kotlin sources, but it can't find the kotlin sources in
PermissionController because they are all in a filegroup.  Add it
manually as a workaround.

Bug: 225013372
Test: m checkbuild
Change-Id: If6b909f336ecafd07d3c2b82f7b2f79ea186b59e
Merged-In: If6b909f336ecafd07d3c2b82f7b2f79ea186b59e