Fix OTP flag removal not persisted for non-OTP background permissions
In PermissionControllerServiceImpl.onOneTimePermissionSessionTimeout, it
is possible for a permission group that does not support one-time
permissions to have the OTP flag. This happens when this permission is
self-revoked. In this case, the flag is removed at the end of the
session to have a consistent UI on the permission settings page.
Previously, this change was not persisted for background permissions.
Currently, every background permission belongs to a group which supports
OTP but this behavior would have lead to UI inconsistency if such permissions
are added in the future.

Test: run cts -m CtsPermissionTestCases -t
android.permission.cts.RevokeSelfPermissionTest
Bug: 225386829

Change-Id: Iab386dc8f944209c2b6cc8ed1b2d28976836c9a1