Check status severity level against max severity level
We decided that we are going to allow sources to send a status severity
level higher than the issues severity levels. This is to cope with
sources that keep track of dismissed issues internally. As a consequence
we must block sources from abusing the status severity level.

TODO(b/223366949): Test these validation checks

Test: atest CtsSafetyCenterTestCases
Bug: 219700241
Change-Id: Ifb2a109da8f56ab1904ebc111665bededf17623c