Remove max line limit from footer summary
Under larger font sizes, we were reaching the max line limit for the
footer summary on the "Unused apps" page which would cut it off. As the
footer is several lines, we use the SettingsLib FooterPreference
instead of the default preference class. This also moves the icon to
the top which aligns better with the other UX in Settings

Bug: 220584404
Test: reproduce with different font sizes, line is not cut off
Change-Id: I856d2e3e00ff873a85609625070f6ed81e77ff98