Guard against null values in SafetyCenterService.
It's already checked by SafetyCenterManager, but technically it's still
possible to pass null values by calling the AIDL interface directly.

We recently had to debug an issue in a local CL; where a test was
failing due to a faulty Parcel implementation that was causing the
package name to be null.
These types of checks will make it easier to debug similar issues
in the future.

Bug: 223216332
Test: atest CtsSafetyCenterTestCases
Change-Id: I7451d392511429d8eb669f912f876c55ce046472