Show custom allow/deny warnings for media perms
When a user manually allows or denies a media storage permission group
(READ_MEDIA_AURAL or READ_MEDIA_VISUAL), we want to warn them and give
them a chance to cancel.

We already show warning dialogs for certain permission group changes,
but the current warning dialog code:

1. Only handles the "deny" case
2. Doesn't allow custom icons, titles, or theme
3. Has hardcoded button handlers (and we need different handling)
4. Is overly verbose

This commit introduces a new warning dialog function to satisfy
these needs, and invokes it when a media storage permission group is
changed.

NOTE: this new warning dialog function could also be used for the
existing warning dialogs. But, for the sake of not causing regressions
or extra test cases, we leave the old dialogs alone for now. This
should be revisited later.

Bug: 225321456
Test: manual
Change-Id: Ibd94b74cb9f0caf0c5a09dd75b68573b3c26c636