Implement SafetyCenter ids.
Essentially an updated version of ag/17017505.

This is extracted in a separate library so that it can potentially be
used by UI as well (it's not ideal, the UI shouldn't have to care about
these IDs; but we will likely have to use it to e.g. be able to group
issues by sources — hopefully we will get rid of this by updating our
APIs in U).

As discussed on ag/17017505 we're ok with the size increase due to using
protos, because we have to use protos to persist dismissed issues ids.

Bug: 217406624
Test: atest CtsSafetyCenterTestCases
Change-Id: I87aad95f9a4acc4f24788363f84c928d3fad007b