Allow not overriding user decision for app op permissions.
When user denies an app op permission in Special app access, it's
actually not reset back to the default mode, but set to modes like
MODE_ERRORED instead. So this can serve as a hint for user-set status
and we can respect it during reconciliation.

Fixes: 221080814
Test: presubmit
Change-Id: I32adc38d61b8b9ba8688ae6c5abb9e356ddcd120
Merged-In: I32adc38d61b8b9ba8688ae6c5abb9e356ddcd120