Fix HibernationSettingStateLiveData NPE
NPE is because the package is uninstalled between the function
suspending and returning, causing the LightPackageInfo value to be null.
We instead store the value when it is non-null and use that throughout.

Stale package data is fine here as any API needs to robustly deal with
the package not being installed.

Bug: 227861199
Test: builds
Change-Id: I892fab5707816f540b3effaf6e725011cde3f66d