Add the infrastructure to allow OEMs to make small modifications
It can be expensive for OEMs to overlay the whole XML config.
Make every field in the XML config accept string references.
Remove some prohibited checks so that we can add placeholder values that
can be overlayed later on if the source state is also overlayed.
Fix flaky overlay test as best as we can.
Fix linter to be able to read string references.
Add missing presubmit test mappings.

Test: m out/soong/.intermediates/packages/modules/Permission/SafetyCenter/Resources/SafetyCenterResources/android_common/lint/lint-report.html
Test: atest CtsSafetyCenterTestCases
Test: atest SafetyCenterConfigTests
Test: atest ConfigLintCheckerTest --host
Bug: 229192140
Change-Id: If8fa41db9d2424ca5a311338e88cb9bb92a632b8