Do not auto-revoke if permission changed recently
Introduce PermissionChangeStorageImpl which stores the day when a
package's permissions are changed by the user. This data is retained for
90 days.

When we auto-revoke, check if the permissions were changed within the
last 90 days and if so, do not auto-revoke any of the package
permissions.

For CTS testing coverage, we also add a setting to allow the storage to
store exact times (so we aren't stuck waiting days in the CTS test)

Bug: 205642821
Test: atest PermissionChangeStorageImplTest
Test: atest AutoRevokeTest
Test: manual
Change-Id: I7866a1209bcb388d8efa6522cad9e57b280b1996