Update the storage permission dialog to match system ui dialogs
This doesn't quite meet the specs of system ui dialogs since a lot of
resources are private, fixing the colors is impossible unless the colors
are exposed as system api. The layout differences could in theory be
fixed but this would require reimplementing the alert dialog in the
permission controller.

Test: Manual
Fixes: 225321456
Change-Id: I968a0e669bf2a49f245b6b0c3a6b2abe3b84d2ce