Create a special case for the device lock group summary
Programmatically use the titles of the sources in the device lock group
as the summary of the group. This is to simplify the life of OEMs that
do not want to create an overlay for each device variant that has
different device lock capabilities. We should consider making this less
hacky in U by adding a formal API for this.

Bug: 229192140
Test: manual
Change-Id: I3e911477c51b6407274dd82dcd35be84552ae6bd