Add READ_MEDIA_* permission to storage role
With SDK 33, the READ_EXTERNAL_STOARAGE and WRITE_EXTERNAL_STORAGE
permission become obsolete. But, the "storage" permission set currently
only grants these permissions, so apps with this role will lose storage
access with SDK 33. The solution is to add the READ_MEDIA_* to this
permission set, because these permissions now provide the access that
READ_* used to provide.

Bug: 229150830
Test: atest DefaultPermissionGrantPolicyTest RoleParserTest
Change-Id: I452432d072ba60bbf19f7f66916d35bd1b3923ed