Make revokeAppPermission split-perm aware
AutoRevoke has a special set of rules to determine whether or not a
particular permission group is revocable or not; only revocable
permission groups will be revoked.

However, we currently fail to properly account for split permissions.
If permGroup A splits into B, then, we need to always either revoke
both A and B, or revoke neither A nor B. If we don't, we leave them
in an inconsistent state.

This commit adds the extra checks necessary to ensure that split
permissions are handled properly.

Bug: 223660952
Test: atest AutoRevokeTest
Change-Id: I30df47c4e0f81637d4d39eb200548c90a3df9409