Fix unused preference recycling
The data object was holding on to the view holder object in the form of
an ImageButton. When this view holder was recycled, that view button
could be set again by the old data object.

The right way to do this is to not store any view data in the data
object since that breaks the recycling contract.

Bug: 223664033
Test: repro from bug
Change-Id: Idb35fcc72e9b3322a79517a9a45852277ca2567a