Replace gensrcs with java_library to deprecate depfile in gensrcs
The depfile concept in Soong and Ninja isn't applicable in Bazel because Bazel requires listing the deps explicitly. This CL changes the module type from gensrcs to java_library with proto.type set to "stream".

The end goal is to eventually deprecate depfile in gensrcs to ensure all gensrcs modules are convertable to Bazel.

Test: CI
Bug: 179452413
Change-Id: Icd588b33749a423ed2518d4c3a51909e962f36e0