Conditionally hide 'Files' permission group
On the T platform, the "Files" permission group currently only affects
apps with legacy storage access. So, hide "Files" for apps without
legacy storage access. (And, only do this on a T+ platform.)

Apps have access to legacy storage when either:

- targetSdk<Q, or
- targetSdk==Q, and OPSTR_LEGACY_STORAGE app op is allowed

This affects the following screens in PermissionController:

- ManageStandardPermissionsFragment
- PermissionAppsFragment
- AppPermissionGroupsFragment
- AllAppPermissionsFragment
- AppPermissionFragment

Also: hide the subtitles that appears on the "Files" permission group for
apps with non-legacy-storage.

Change-Id: I6d10b96bd5540d994fca487dd8f06a3780cd245e
Bug: 235277006
Test: manual