Only dispatch the implicit ENABLED_CHANGED broadcast to relevant users.
It doesn't seem necessary to dispatch it for all users.

The explicit broadcast should only send to users the sources are
interested about, e.g. if a package only has a primary user source then
it shouldn't receive the enabled changed broadcast in the work profile.

The implicit broadcast is only used by persistent processes to enable /
disable a state, so only dispatch to system user.
This also matches the documentation which specifies that it is sent
implicitely "system-wide".

Bug: 235829956
Bug: 203098016
Test: atest CtsSafetyCenterTestCases
Change-Id: Ibaeb6d317cfcd80369f19fa0aae206fa6dcf4a7c