Allow wait for broadcast idle to timeout.
This can cause some tests to timeout after waiting for the broadcast
queue to be idle for more than 60 seconds. Limit this waiting to 10
seconds and allow it to timeout, hoping for our broadcasts to get
through in that time.

Also wait for broadcast idle for tests that don't hold the broadcast
receiver permission: this could cause a flake where we weren't waiting
long enough for the broadcast to be dispatched (and dismissed due to not
having the permission), leaking onto other tests.

Bug: 228426002
Test: atest CtsSafetyCenterTestCases
Change-Id: I08eada348c399213f486ea48be24458821d6d0fd