Fix StoragePermissionAppsFragmentTest
This test was broken by a recent change to the UI.

Apps with targetSdk>=30 are now hidden from the "Files" permission
group screen (and apps with targetSdk==29 are conditionally hidden).

But, the test uses a helper app with targetSdk==30, and happens to rely
on the visibility of the exact element which is now hidden.

The solution used here is to change the test helper app to
targetSdk==28, which retains old behavior, and shouldn't affect the
intention of the test in any way.

Change-Id: I03f24bbfe2cb9099dbf184ad2eb866d102d8979d
Fix: 235638067
Bug: 235277006
Test: atest StoragePermissionAppsFragmentTest