Clear Binder identity before calling createPackageContext
This is likely what ends up causing the crash in b/236006800 and b/232929818. This shouldn’t be necessary but should hopefully fix the test failure for now while we look into whether this is highlighting a real problem that needs to be fixed elsewhere.

Bug: 236006800
Change-Id: Ib4ca63f5490653506c2056274851d635d007acaf
Merged-In: I0374802645fbff51d86417cf492b23d4dc1b1377
Test: atest CtsSafetyCenterTestCases