Add auto-revoked apps as a privacy source
We add whether apps were auto-revoked recently as a privacy source. The
only real call to action here is to review the unused apps, and if the
user goes to the page, the card is removed.

We also persist whether the user needs to review the unused apps and
clear this when the user does visit the page. This ensures the card
persists across SC being closed / device reboots

Bug: 224846118
Test: atest AutoRevokeTest
Change-Id: Iff61f9a31449a9dfc0f04ad36d1f79c47ff268ae