Brand unused apps notif w/ PbA branding
Brand the "X unused apps" notification with safety center branding. This
includes the "Protected by Android" app title and icon. We do not intent
to safety center directly as there is no call to action besides just
reviewing the app and instead keep the intent to go to the "Unused Apps"
page.

We will add a card in a separate CL with a CTA to review unused apps.

Bug: 224846118
Test: manual, see screenshot in bug
Change-Id: Ic1b9064705a2407a81272f7682225f8f90c4bbe2