Wait for broadcast idle before running tests.
That should avoid us having to timeout this call, since the problem
seems to be that BOOT_COMPLETED broadcasts are being dispatched prior to
running our tests which then causes `waitForBroadcastIdle()` to run for
too long.

Hopefully the RunCommandTargetPreparer is allowed to run for long
enough, which should ensure all further calls to `waitForBroadcastIdle()` complete on time.

Bug: 237104364
Test: atest CtsSafetyCenterTestCases
Change-Id: I4a57c5892a7835eea4e1da0d5bb5f6eb36678ea1