Implement Safety Center error entries.
As per latest UX conversations:
* Clear data and show different summary for sources that have
  errored-out (unless it's an inline action error)
* Clear errors prior to starting a refresh
* Use the "review your settings" string for the overall status summary
  when there is at least one error
* Remove the IFTTT and instead rely on strings equality; it's not the
  greatest approach but it seems better than duplicating the logic

We still need to mark sources that timeout as having errored-out; but
that's a bigger effort since it will likely break the UI tests. We also
need a different string for the error title. I'll follow up
for that with other CLs. In the meantime this will allow us to resolve
the SecurityCheckup problems (showing stale data when an error was
reported).

Bug: 237430874
Bug: 237057234
Test: atest CtsSafetyCenterTestCases
Change-Id: Ifa99a26523b78e992eb31e53c76c31fe9776ccb6