Use UUIDv4 (pseudorandom) for refresh broadcast IDs
I don't think the suggestion in the bug to encode other data about the
refresh request inside the broadcast ID is useful or a good idea.

It looks like there is a possibility of collisions with the existing
implementation so I'm proposing this change to use UUIDs instead.

When testing I see the expected logcat output like

Starting a new refresh with refreshReason:100 \
  refreshBroadcastId:e800f468-c55c-4d8a-92ed-1efeb4625275

Test: Manually checked refreshes work as before
Bug: 223120199
Change-Id: Id36561335095f291f84cc28ef9d5cf60c2de898a