Only send broadcasts which will resolve
Also changes the way the refresh tracking works so that sources are not set
"in flight" if no broadcast was sent to them.

To make the above change the SafetyCenterBroadcastDispatcher was
brought under the same mApiLock as the SafetyCenterRefreshTracker to
ensure thread-safety, although further changes may be made in this
area to reduce contention on that lock.

Test: Manually tested by adb uninstalling com.android.vending
Bug: 235222142
Change-Id: I557c1fcb77e2a998e5d729859f5d8546ee157e5a