Send data at end of scan instead of throughout.
A scan == user clicks rescan button. We can pause all updates until the
scan is finished since the rescan button is only shown when the status
is green and there are no issues. (a previous version of this CL paused
updates for data loading aka page open scan, and allowed updates through
for interactions with issues present on the page.)

This also adds some logic to debounce/skip irrelevant data updates
received by the UI while processing the queue.

Bug: 239809022
Test: Manual, CTS tests incoming
Change-Id: I682904908256856f00803bdc955094f3d80f97da