Use OVERALL_SEVERITY_LEVEL_UNKNOWN where applicable.
This should ensure that we don't log "looks good" when in fact we don't
know the status.

The UI already handles this status well (and behaves just like a green
state in that case).

I've also made some tweaks to the strings returned, in particular to
avoid saying "Looks good" to the user when we don't know yet.

If the status is unknown (i.e. we don't have a yellow or red issue, and
at least one entry doesn't have data / has errored-out):

* Show "Review settings" if no scan is in progress
* Show "Loading" if a page open scan is in progress

Bug: 242054626
Bug: 238319910
Test: atest SafetyCenterManagerTest
Change-Id: Iddab7e9b06240a092277d30a6083469f6b1f2da4