Control visibility manually for issue subtitle and PbA widget.
The previous constraint layout group would unexpectedly override the
visibility of any children. So setting the visibility to GONE of a view
inside a Group would have no effect.

This is supposedly fixed in some capacity in android's checked-in
version of ConstraintLayout (v2.0 beta 7, originally fixed in v2.0
beta6) but the behavior's broken and we don't really have the time
or willpower to update android to use a newer version of
ConstraintLayout.

See the bug for more details / complaints.

Bug: 242705351
Test: atest CtsSafetyCenterTestCases:SafetyCenterActivityTest, also
verified manually there are no visual bugs when rendering different
issues or playing issue resolution.

Change-Id: Ia273cd984e97ede64a00fcdf9d18fd727a97048a