Fix a slow CTS test.
This CTS test was doing two things at the same time: wait for the
success message to show and then wait for the issue to be removed.
However, since we use a large timeout for the animation in order to show
the success message for long enough, it meant that the second assertion
would wait for a very long time. Splitting this into two tests (one that
overrides the animation timeout, one that doesn't), resolves this
problem.

Bug: 244708990
Test: atest CtsSafetyCenterTestCases
Change-Id: Ib04ed05858804e1ab5b2a34e31498d49c9d38ca5