Fix permanent loading UI
Loading UI could permanently stay on screen due to a race where the data
could load in after we check if data is loaded but before we set the
loading state.

This CL makes these operations atomic by ensuring the check also happens
on main thread before modifying the view.

Bug: 244219949
Test: manual
Test: atest AppHibernationIntegrationTest

Change-Id: Ib6290ea15d014f6bf6ec7d19a7fff74228485a56