Handle StaleObjectException when clicking invalid items.
It's possible that the UI listener gets called with new data after we
found a UI element to click, which could cause a StaleObjectException if
this element is getting replaced.

Allow retrying in this case by handling the StaleObjectException and
getting the potentially updated UI element to click on it again.

Bug: 221401739
Test: atest CtsSafetyCenterTestCases
Change-Id: I0664ec08d63b4be01fa8f794147b67839d64a990