Make dismissed issue counts more consistent
I think this is a non-trivial, behavioural change which fixes a subtle
bug in how the dismissed issue counts logged to Westworld are
computed.

In `writeSafetySourceStateCollectedAtom` an issue was counted as
dismissed if it had ever been given a non-null `dismissedAt` time, but
now the `isDismissed` logic is used which accounts for resurfaced
issues.

In `pullOverallSafetyStateAtom` there was a similar bug but I've fixed
it in a different way because the open/dismissed issue counts here are
based on the computed `SafetyCenterData` and not the underlying
`SafetySourceData`.

I was looking at this code with a view to refactoring it, I wanted to
make the code for checking issue dismissals more consistent. The
overall objective is to make it possible to extract all the code to do
with the issue cache to a separate class.

Bug: 223543358
Test: atest SafetyCenterManagerTest
Change-Id: I8ee4ff08f0971a46a56c9cbf38c987560d1d364b