Migrate STORAGE permissions to READ_MEDIA on T
On Android platform T, when the platform was upgraded, the
READ_MEDIA_* permissions were not always getting auto-granted.

Specifically: when an app has READ_EXTERNAL_STORAGE granted, but also
has {READ_MEDIA_IMAGES and/or READ_MEDIA_VIDEO and/or READ_MEDIA_AUDIO}
in its manifest at the time of the platform upgrade. When the platform
upgrade occurs: whichever of the READ_MEDIA_ permissions are in the
manifest will not be granted, and vice versa (whichever of the
READ_MEDIA_ permissions are not in the manifest will be granted.)

But, if these permissions are in the manifest, then we want them to be
auto-granted if (and only if) the STORAGE permissions
(READ_EXTERNAL_STORAGE, WRITE_EXTERNAL_STORAGE) were granted.

The solution used here is to perform a one-time upgrade patch which does
the following (upon upgrading to the T platform, for all apps):

1. If the user has previously granted the app  READ_EXTERNAL_STORAGE or
   WRITE_EXTERNAL_STORAGE, then grant all READ_MEDIA_* permissions which
   are requested in the manifest
2. Regardless of #1, revoke READ_EXTERNAL_STORAGE and
   WRITE_EXTERNAL_STORAGE

Fixes: 249702391
Test: atest RuntimePermissionsUpgradeControllerTest
Change-Id: Ieb048f5d38331141bb96b2a32d94e81233a5465e