Revising names in SafetyCenterIssueCache.
This is a follow-up on ag/20074168 where I was trying to keep the diff
as small as possible. Now we have the issue cache stuff in a separate
file lots of names can be shortened or otherwise clarified.

I also moved some snippets of code for transforming to/from
PersistedSafetyCenterIssues into the IssueData class.

Bug: 249979479
Test: atest SafetyCenterManagerTest
Change-Id: Ia64f9f2bbbc6b3fbba3ccee4b309e14639912698