Stop passing SafetyCenterRepository calls through SafetyCenterDataTracker
This CL properly separates the API of the new "repository" and the
"tracker" which will be renamed in a subsequent CL (b/250812300).

Although the tracker is still over 1k lines, its public API is now
just three methods (getSafetyCenterData, getDefaultSafetyCenterData
and pullAndWriteAtoms).

The SafetyCenterRepository now holds all the state and its API includes
all the methods to read and mutate that like dismissSafetyCenterIssue
and clearForUser.

Fix: 250813415
Bug: 223543358
Change-Id: Iad2e66e0a55c7ac69335cc61bb4b49dab21bc6f3