Handle clicking outside of AlertDialog
This AlertDialog has two buttons, "Back" and "Confirm". But, there's
actually another action that can occur: the user can click outside the
dialog. Currently, we don't handle this action, and the default behavior
is undesirable (it behaves like "Confirm").

Change the behavior of clicking outside the dialog to match the behavior
of clicking "Back".

Test: atest MediaPermissionTest
Bug: 244144105
Change-Id: I20ba44edd4144471b636ff26831a43e3bcb7a2ec