Fix an update deduping problem in SafetyCenterListeners.
We were pushing the very first update on the original
IOnSafetyCenterDataChangedListener and not on the wrapper, which means
the first update was not updating the mLastSafetyCenterData; making it
possible to push the same SafetyCenterData without any change (so no
"update").

Technically it's hard to repro this situation in real life because the
API doesn't actually push any data to listeners if it believes it won't
change the underlying SafetyCenterData (e.g. a source pushing the same
data).

And even if it does happen, this seems relatively unimportant as
listeners can handle duplicated SafetyCenterData.

Bug: 254254431
Test: atest CtsSafetyCenterTestCases
Change-Id: I32456ee7d349f1f9ed4353eabaf5ecae30e985b7