Safety Center API changes for notifications
Introduces a new class SafetySourceIssue.Notification which may
optionally be added to any SafetySourceIssue to specify certain
properties of a Notification for that issue.

Also adds a new NotificationBehavior @IntDef and property of
this type on the issue model.

CL includes CTS tests for these model APIs.

Please see go/angel-notifications-2023 for more context.

Test: atest SafetySourceIssueTest
Fix: 255945913
Change-Id: Icee490a73d67afbe92e30f5fbe686bc564a1c062