Load storage/media perm group icons from fb
All permission groups have icons, which are usually stored in
frameworks/base as resources. In Android T, we created a few new
UI flows which show the icons for the STORAGE, READ_MEDIA_VISUAL,
and READ_MEDIA_AURAL permission groups in alert/grant dialogs. But,
rather than loading the icons from frameworks/base, we copied them
into PermissionController and used them there. Now, we've realized
this is undesirable for OEMs, because (1) we failed to make them
overlayable, and (2) it's confusing to OEMS that we have two copies
of the icons.

This change ensure all of these icons are loaded from frameworks/base.

Bug: 254673246
Test: atest MediaPermissionTest; manually verify icons shown on screen
Change-Id: Ief8b3bb016dab302a8bc8f29844ec75ae9b9995e