Use proper count for total number of apps
As part of the permission decluttering effort, NFFs forked logic for how
we count and displays apps that use permissions on permission manager.
This was originally only counting apps where the user explicitly set the
permission, which was useful for filtering permission categories, but
not for depicting the total number of permissions.

Bug: 242662819
Test: Manual, check total number of apps on permission manager
Change-Id: I628cb81ed37a208fed9386e0f7e8aac358f98c7a