Move and improve collection classes.
Move collection classes into a collection package so that it can be
imported with wildcard, while we put other utility classes in the util
package.

Added indexed methods for List so that we can still do optimized
indexed for loops instead of using iterators.

Added noneIndexed() to match allIndexed() and anyIndexed().

Improved getWithDefault() and putWithDefault() for our typical
usage. Now getWithDefault() accepts a nullable this, while
putWithDefault() returns whether a change has happened.

Reordered methods to be sorted alphabetically.

Also added two extension functions on Int for working with flags.

Bug: 182523293
Test: presubmit
Change-Id: Ie03d9ec7088b1980615284c9ca06ee7351f5b5aa