Improve access decision listener.
Make it lock-free for reads and deliver callbacks synchronously, so
that as long as compatibility layer aggregates the callbacks in a fast
and reasonable way, we can better deal with large scale changes.

Also make putWithDefault() return the previous value instead of
whether any change happened, because we may need the previous value in
certain cases (e.g. notifyOpChangedSync()).

Bug: 182523293
Test: presubmit
Change-Id: I2e9ccabbd40602ca52545e3b6ebb1086707811f2