Remove ConfirmationActivity ActionBar
The theme of this Activity is not a NoDisplay theme, because the
AlertDialog it creates needs to have themeing matching the rest of the
system.

We don't want the Activity underneath to display a titlein case it
accidentally gets revealed during an interaction with multiple dialogs
at once, such as when unblocking the mic during a permission grant.

See change I312976e845b5ed06b952da6eb37e764040174bb4 for history on the
theme, when it first changed from NoDisplay to Alert.

Test: Visual inspection on cf phone and cf tv.
Fix: 227277251
Change-Id: Iad2ddb47ddcbf9784ebcc94d25b4bf3213f1f5e4