Handle notification dismissals
Filters out those notifications which were previously dismissed.

Adds a new BroadcastReceiver class which handles notifications
"delete" PendingIntents for Safety Center notifications. These PIs are
set on the Notification.Builder and sent when a notification is swiped
(or otherwise dismissed) from the notification shade. We record this
dismissal to the issue cache so that no notification will be shown for
that issue (for a while).

Test: Manually
Bug: 255946874
Change-Id: I71703b84b51f24b89b56574e3c3a724f4c926a69