Make executeIssueActionInternal method
This CL moves some code from the Stub inner class to the
SafetyCenterService itself and creates and new
executeIssueActionInternal method so that other code without
SC/system server can execute actions in a consistent way.

Concretely I plan to make SafetyCenterNotificationReceiver call this
function when an "action" button on a notification is clicked.

Test: atest SafetyCenterManagerTest
Bug: 260084296
Change-Id: Ib6ccea28bc0c0d67acb239521cafc29b8d5ce596