Fix SafetyCenterNotificationTest flakes
These tests had a (low) flake % due to a bug in the code of
CtsNotificationListener. Some notifications were posted/removed and
emitted, but not observed by subscribers to the eventFlow.

That helper class has been refactored to use a List rather than a Flow
which has simplified and deflaked it.

Relnote: Deflake Safety Center CTS tests
Test: atest SafetyCenterNotificationTest --rerun-until-failure 100
Bug: 261978453
Change-Id: I97f973f5c3abb619197673a004ed4f09f58277a5