Create SafetyCenterNotificationChannels
Adds a new class called SafetyCenterNotificationChannels which is
responsible for creating notification channels for SC, removing the
existing code for that from the NotificationFactory, in preparation
for expanding it and adding complexity in further CL(s).

Single example of using a string resource via the
SafetyCenterResourcesContext class. But this CL should be merged
simultaneously with the child CL to avoid unnecessary translations.

Bug: 259398016
Bug: 259399024
Test: atest SafetyCenterNotificationTest
Relnote: N/A
Change-Id: Ia9d7cedc1c1c8cd71ed34c814e5b06f8deeae3f6