Fix crash in multi-users.
Spotted thanks to a flaky failure, which happens with a different stack
trace because:

1. A crash happens when calling the API while delivering the data to
   listeners
2. Since this happens in the middle of a listeners dispatch, this causes
   the state to be corrupted for further calls and creates the stack
   trace seen in the bug

On top of fixing this crash, we also ensure that if any crash occurs
while delivering a listener update, the state isn't corrupted. This
should avoid completely breaking Safety Center in case of a crash
while delivering a listener update, on top of making further issues
easier to debug (since we won't be polluted by this stack trace instead
of the real cause for the problem).

This bug should only present on the master branch. The tm-mainline-prod
branch uses PackageUtils.queryUnfilteredIntentActivitiesAsUser which
already clears identity.

Bug: 264718767
Test: atest CtsSafetyCenterTestCases
Change-Id: Ic88d132bdec2cadf3919388e34a44e7a5213fcec