Notification channels in tests
This CL plumbs the resolved NotificationChannels for each
StatusBarNotification so that tests can be written involving the
information contained in those, such as the "importance".

This CL is written to avoid too much churn in the existing
notification tests and new test cases will be added specifically to do
with the multiple notification channels later.

Bug: 264369469
Test: atest SafetyCenterNotificationTest
Relnote: N/A
Change-Id: Id8f49603d80c0ae17d3e2a86ce6a7b00faa6b8ed