Check if intent resolves in package first.
The intent will always be considered to resolve even for implicit
internal non-exported intents, so instead check if the intent resolves
in the package first to make it explicit.

This assumes that the package intent has the highest priority for the
intent action, which is likely intended for the source anyway.

Bug: 265953791
Test: atest CtsSafetyCenterTestCases
Change-Id: Ic2fac1808bdbe469d3daa51785e46ac2b2b94521