Allow hyphenation on our TextViews
Hyphenation used to be on by default, then got turned off because of
performance considerations. Android 13 added 'fast' hyphenation
options that are suitable for all but the most extremely performance
sensitive situations.  But these are not on by default.

Enable these by default for all of our TextViews and similar (buttons).
Our localized strings mostly don't have soft hyphenation yet, but this
change enables translators to take advantage of this.

Bug: 265756249
Bug: 261995127
Bug: 261994180
Bug: 261994733
Bug: 261994555
Bug: 261994847
Bug: 261994846
Bug: 261994512
Bug: 261994377
Bug: 261994548
Bug: 261994514
Bug: 261994513
Test: manual, with test string changed to add soft hyphen
Test: atest CtsSafetyCenterTestCases SafetyCenterConfigTests

Change-Id: I0979bc23d3d51fbb61c3f629abf723930cdb368c