Don't unnecessarily query package metadata for SafetyLabels
Only querying metadata for packages in SafetyLabel display when we know
that the package has permissions that should include safety labels.
Right now that is limited to LOCATION permissions.

As part of this, I consolidated some of the utilities and made it so
there is less branching and less need to check in multiple places
whether the feature is enabled.

Bug: 260873483
Test: atest
Change-Id: I30efd202985fecb95eb919952deaf555d289fc0e