Handle ACTION_MANAGE_PERMISSION_USAGE on Auto
This change adds support to handle the
ACTION_MANAGE_PERMISSION_USAGE subattribution flow on Auto, mirroring
the phone change in Ie6a52873b295d7f202655f6b1a5b246202221d0e.

Ideally this logic should be shared in a ViewModel, but that is
more complicated right now due to ongoing refactors and clean-ups. For
now this code lives within the UI layer, which will also make it easier
to cherry-pick, if necessary.

Bug: 256173612
Test: atest PrivacyDashboardSubattributionsTest
Change-Id: I8f97be2b99014469a00933af7269f56691cba7d7