Split jarjar rules for framework-permission-s and service-permission.
Previously we used the same jarjar rules for both in the hope that if
we ever need to pass objects between them we won't run into class name
issues. However this created an issue when a static lib dependency is
needed in both framework-permission-s and service-permission, because
if we add it to service-permission it will conflict with the copy in
framework-permission-s, but if we don't service-permission will be
missing it on R.

So we should simply decouple the two and jarjar their classes into
different packages, which will help us solve this problem.

Bug: 253568736
Bug: 196909329
Test: atest StrictJavaPackagesTest
Change-Id: I107ac7be4033ccd5b6fb18f73745ded3a5e79202