Use icon for unspecified group severity level.
This originally had an icon in ag/17401037 but this was reverted in
ag/17504366. IIRC this was based on a discussion with deweytyl@ about
how the UI should behave in this case.

I don't think this behavior is desirable anymore so I'm changing this
back to having an icon when the severity is not speficified (see
attached bug for more context); but please shout if you disagree.

Bug: 269097554
Test: atest SafetyCenterFunctionalTestCases
Change-Id: I6228833b3ae3453a5c7a29f8d12950ad024b2eb0